Recover add floor float guard

+++ src/map.c
@@ -80,7 +80,7 @@ void add_floor(Tmap *m)
             }
             else {
                 width=rand();
-                if ((int)(((float)(300-m->room[31].level/5)/300.0f)*10.0f)<1) width=6;
+                if (((float)(300-m->room[31].level/5)/300.0f)*10.0f<1.0f) width=6;
                 else width=width%(int)(((float)(300-m->room[31].level/5)/300.0f)*10.0f)+6;
             }
         }
